medium_clone/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-autoslug==1.9.8
django-crispy-forms==2.0
django-extensions==3.2.1
django-tinymce==3.5.0
easy-thumbnails==2.8.5
Pillow==9.4.0
python-slugify==8.0.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
